requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.3
cryptography==41.0.3
argon2-cffi==23.1.0
python-dotenv==1.0.0

